requirements.txt
feedparser>=6.0.10
loguru>=0.7.0
pyyaml>=6.0
openai>=1.0.0
requests>=2.31.0